backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
transformers
torch
Pillow
tensorflow